requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.16.0
webdriver-manager>=4.0.1
pyyaml>=6.0
schedule>=1.2.0
python-dotenv>=1.0.0
lxml>=4.9.0
ebooklib>=0.18
